@@ -1,36 +1,34 @@
 {
    "private": true,
    "dependencies": {
-      "@babel/cli": "7.10.1",
-      "@babel/plugin-proposal-class-properties": "7.10.1",
-      "@babel/preset-env": "7.10.1",
-      "@babel/preset-react": "7.10.1",
-      "@material-ui/core": "^4.12.2",
-      "@material-ui/icons": "^4.11.2",
-      "axios": "^0.21.1",
+      "@babel/cli": "^7.18.10",
+      "@babel/plugin-proposal-class-properties": "^7.18.6",
+      "@babel/preset-env": "^7.18.10",
+      "@babel/preset-react": "^7.18.6",
+      "@material-ui/core": "^4.12.4",
+      "@material-ui/icons": "^4.11.3",
+      "axios": "^0.27.2",
       "classnames": "^2.3.1",
-      "connected-react-router": "^6.9.1",
-      "fuse.js": "^6.4.6",
-      "history": "^4.7.2",
-      "howler": "^2.2.1",
+      "connected-react-router": "^6.9.3",
+      "fuse.js": "^6.6.2",
+      "history": "^5.0.0",
+      "howler": "^2.2.3",
       "mousetrap": "^1.6.5",
-      "node-sass": "^4.14.1",
-      "nouislider": "^14.7.0",
-      "prop-types": "^15.7.2",
-      "react": "^17.0.2",
-      "react-dnd": "^11.1.3",
-      "react-dnd-html5-backend": "^11.1.3",
-      "react-dom": "^17.0.2",
-      "react-icons": "^4.2.0",
-      "react-redux": "^7.2.3",
-      "react-router-dom": "^5.2.0",
-      "react-scripts": "3.4.1",
-      "react-select-search": "^3.0.5",
-      "react-slick": "^0.28.1",
-      "redux": "^4.0.5",
-      "redux-thunk": "^2.3.0",
+      "node-sass": "^7.0.1",
+      "prop-types": "^15.8.1",
+      "react": "^18.2.0",
+      "react-dnd": "^13.1.0",
+      "react-dnd-html5-backend": "^12.1.0",
+      "react-dom": "^18.2.0",
+      "react-icons": "^4.4.0",
+      "react-redux": "^8.0.2",
+      "react-router-dom": "^5.3.3",
+      "react-scripts": "4.0.3",
+      "react-select-search": "^3.0.10",
+      "redux": "^4.2.0",
+      "redux-thunk": "^2.4.1",
       "shared": "file:../shared",
-      "use-force-update": "^1.0.7"
+      "use-force-update": "^1.0.10"
    },
    "scripts": {
       "start": "react-scripts start",
