@@ -16,7 +16,8 @@ class CardForm extends Component {
 }
 
 CardForm.propTypes = {
-   classes: PropTypes.object.isRequired
+   classes: PropTypes.object.isRequired,
+   children: PropTypes.oneOfType([PropTypes.element, PropTypes.array]).isRequired
 };
 
 export default CardForm;
