@@ -1,5 +1,12 @@
 import React from "react";
+import PropTypes from "prop-types";
 
-export default function ButtonRow({ children }) {
+function ButtonRow({ children }) {
    return <div style={{ width: "100%", display: "flex", justifyContent: "space-around" }}>{children}</div>;
 }
+
+ButtonRow.propTypes = {
+   children: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)]).isRequired
+};
+
+export default ButtonRow;
