@@ -47,7 +47,8 @@ function getSizingValue() {
 ResizableContainer.propTypes = {
    classes: PropTypes.object.isRequired,
    isGame: PropTypes.bool,
-   noToken: PropTypes.bool
+   noToken: PropTypes.bool,
+   children: PropTypes.oneOfType([PropTypes.element, PropTypes.arrayOf(PropTypes.element)]).isRequired
 };
 
 export default withStyles(styles)(ResizableContainer);
