@@ -331,7 +331,8 @@ class CardSearch extends PureComponent {
 }
 
 CardSearch.propTypes = {
-   classes: PropTypes.object.isRequired
+   classes: PropTypes.object.isRequired,
+   newResults: PropTypes.func.isRequired
 };
 
 export default connect(null, { newResults })(withStyles(styles)(CardSearch));
